Enable repo implementations to save storage.
